move cardTuples to src/functions
