+++ src/classes/index.ts
@@ -2,8 +2,5 @@ export * from './AppContext';
 export * from "./Player";
 export * from "./PlayerStore";
 export * from "./PlayingCard";
-export * from "./deprecated/PlayingCardDeck";
 export * from "./PlayingCardSort";
 export * from "./PlayingCardSuit";
-export * from "./cardTuples";
-
