+++ src/functions/cardTuples.ts
